Force KuCoin clients to use IPv4 transport
